Regex's were running a bit slow. Well it's because this library uses [regexp2](https://github.com/dlclark/regexp2) which implements look arounds, which go's `regexp` does not support. Matomo relies on good amount of lookarounds, but this hurts performance.

Wouldn't it be nice to use `regexp2` when the regex needs lookarounds, and go's `regexp` when it does not?

This PR abstracts away which regexp implementation we're using so that things can run faster _most_ of the time.

Out of the several benchmark tests I've written, I'm seeing parsing time decrease by ~40%.

Before this pr:
```
λ go test -bench=.
goos: darwin
goarch: amd64
pkg: github.com/muxinc/devicedetector
BenchmarkParseDeviceInfoForBrowser-12                 50          24958820 ns/op
BenchmarkParseDeviceInfoFor-12                        33          35534547 ns/op
BenchmarkParseDeviceInfoForMobile-12                  54          26580086 ns/op
PASS
ok      github.com/muxinc/devicedetector        7.008s
```

After:
```
λ go test -bench=.
goos: darwin
goarch: amd64
pkg: github.com/muxinc/devicedetector
BenchmarkParseDeviceInfoForBrowser-12                 90          13884112 ns/op
BenchmarkParseDeviceInfoFor-12                        58          20383983 ns/op
BenchmarkParseDeviceInfoForMobile-12                  99          13111794 ns/op
PASS
ok      github.com/muxinc/devicedetector        7.244s

```